Fixes the docker compose errors when running npm run env:start
Error:```
ashfame@Ashfames-MacBook-Pro wordpress-develop % npm run env:start

> WordPress@5.9.0 env:start
> node ./tools/local-env/scripts/start.js

WARN[0000] The PHP_FPM_UID variable is not set. Defaulting to a blank string.
WARN[0000] The PHP_FPM_GID variable is not set. Defaulting to a blank string.
WARN[0000] The PHP_FPM_UID variable is not set. Defaulting to a blank string.
WARN[0000] The PHP_FPM_GID variable is not set. Defaulting to a blank string.
WARN[0000] The WP_MULTISITE variable is not set. Defaulting to a blank string.
WARN[0000] The PHP_FPM_UID variable is not set. Defaulting to a blank string.
WARN[0000] The PHP_FPM_GID variable is not set. Defaulting to a blank string.
WARN[0000] The GITHUB_REF variable is not set. Defaulting to a blank string.
WARN[0000] The GITHUB_EVENT_NAME variable is not set. Defaulting to a blank string.
services.phpunit.environment.LOCAL_PHP_MEMCACHED must be a string, number or null
node:child_process:903
    throw err;
    ^

Error: Command failed: docker-compose up -d wordpress-develop
    at checkExecSyncError (node:child_process:826:11)
    at execSync (node:child_process:900:15)
    at Object.<anonymous> (/Users/ashfame/git/wordpress-develop/tools/local-env/scripts/start.js:8:1)
    at Module._compile (node:internal/modules/cjs/loader:1101:14)
    at Object.Module._extensions..js (node:internal/modules/cjs/loader:1153:10)
    at Module.load (node:internal/modules/cjs/loader:981:32)
    at Function.Module._load (node:internal/modules/cjs/loader:822:12)
    at Function.executeUserEntryPoint [as runMain] (node:internal/modules/run_main:79:12)
    at node:internal/main/run_main_module:17:47 {
  status: 15,
  signal: null,
  output: [ null, null, null ],
  pid: 47656,
  stdout: null,
  stderr: null
}
```

Apparently docker compose files doesn't like boolean values as YAML parser converts them, so they must be specified in quotes, and accordingly handle when checking for their value or using them.

The variables defined in .env file raised the same error so I had to comment them out so that their default value could be picked up.

See: https://devops.stackexchange.com/questions/537/why-is-one-not-allowed-to-use-a-boolean-in-a-docker-compose-yml

I should also mention these happened while setting up the repo on Apple M1 Macbook Pro